fix(welllog): Stage 1 review findings D-006/D-007/D-008 (#285)
Remaining lower-confidence findings from the /qt-cpp-review (companion
to #284 which fixed D-001..D-004):

D-006 (conf 62): prepare_for_export performed no mutation — marked const
so a const WellLogSession& can trigger export-density re-prepare, matching
the const prepared_scene() getter it sits beside.

D-007 (conf 65): export_pixel_height was uint32_t at the binding, so a
Python int >= 2^32 truncated silently (shiboken unsigned-long→unsigned-int).
Changed the typesystem to unsigned long long (64-bit, no truncation),
threaded uint64_t through the binding + prepare_for_export, and added an
explicit > UINT32_MAX rejection in prepare_for_export (returns
invalid_viewport) so absurd values surface as an error rather than a
wrong-density scene.

D-008 (conf 60): hoisted first_curve_id_text.toStdString() once per track
(was computed twice — scale-range and scale-unit lookups). Also removed
the now-redundant PyErr_Clear() after dict_get_float (the getters clear
their own errors since the D-002 fix).

Deferred: D-005 (add_pattern&& overload) — would change the public
ScenePresentationBuilder API for a marginal perf gain on a handful of
patterns per export; not worth the API churn.

Verified: 10/10 python qt-embedding; 10/10 C++ exporter+session; 21/21
host engine_bridge+export.

+++ well-log-engine/src/python/numpy_bridge.hpp
@@ -39,7 +39,7 @@ submit_multi_well_section(WellLogView *view, PyObject *payload) noexcept;
 // aggregate density (T3 / #275) for correct fixed-page pagination.
 [[nodiscard]] PyObject *
 export_scene_svg(WellLogView *view, const QString &document_id,
-                 std::uint32_t export_pixel_height = 0) noexcept;
+                 std::uint64_t export_pixel_height = 0) noexcept;
 
 // Render the prepared scene for ``document_id`` to PDF and return the
 // document bytes (T2 / #274). Text is glyph outlines (non-searchable,
@@ -47,7 +47,7 @@ export_scene_svg(WellLogView *view, const QString &document_id,
 // into export-density re-prepare (T3 / #275).
 [[nodiscard]] PyObject *
 export_scene_pdf(WellLogView *view, const QString &document_id,
-                 std::uint32_t export_pixel_height = 0) noexcept;
+                 std::uint64_t export_pixel_height = 0) noexcept;
 
 } // namespace python
 } // namespace welllog
